refactor(tui): split config monolith into clean leaf modules (#3311)
Extract self-contained, behavior-preserving leaves out of the
~6.7k-line crates/tui/src/config.rs into sibling modules under
crates/tui/src/config/, each re-exported behind a `pub use` facade so
every existing `crate::config::<symbol>` path resolves unchanged. No
call site outside config.rs is touched; entangled credential/provider/
default-model/normalization/route logic stays put.

Modules created (config.rs: 6739 -> 6246 lines, ~493 lines moved):
- config/models.rs (160): provider model-name + base-URL constants and
  curated model lists (DEFAULT_*_MODEL/BASE_URL, RECENT_OPENROUTER_*,
  COMMON/OFFICIAL_DEEPSEEK_MODELS, ...). API_KEYRING_SENTINEL stays in
  config.rs with the credential logic.
- config/search.rs (135): SearchProvider, SearchProviderSource,
  SearchProviderResolution, SearchConfig (self-contained [search] types).
- config/subagent_limits.rs (69): sub-agent concurrency/timeout limit
  constants + their two private clamp resolvers (pulled back privately,
  no new external surface).
- config/paths.rs (203): pure filesystem path helpers (config/cache/
  workspace path resolution, env-var path overrides, ~ expansion);
  effective_home_dir/expand_path re-exported pub(crate). Workspace-trust
  and config-load logic stay in config.rs.

scripts/check-provider-registry.py now also scans config/models.rs for
the default model/base-URL constants so the registry gate follows the
split.

Verify: config suite 480 passed / 0 failed; full bin suite 5248 passed
/ 0 failed; cargo build -p codewhale-tui --locked green; clippy clean on
all changed files (the 3 remaining -D warnings errors are pre-existing
too_many_arguments lints in untouched chat.rs/engine.rs); cargo fmt
--all clean; provider registry drift check passes.
